Remove tree hashing from ssz crate
