--- eth2/utils/ssz/src/signed_root.rs
@@ -1,5 +1,0 @@
-use crate::TreeHash;
-
-pub trait SignedRoot: TreeHash {
-    fn signed_root(&self) -> Vec<u8>;
-}
